requirements.txt
# Core dependencies
requests>=2.31.0
python-dateutil>=2.8.2

# AI Providers (optional - install as needed)
anthropic>=0.25.0
openai>=1.30.0
google-generativeai>=0.7.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0